+++ src/main/java/edu/hnu/conference_system/domain/UserInfo.java
@@ -20,8 +20,7 @@ public class UserInfo implements Serializable {
      * 
      */
     @TableId(type = IdType.AUTO)
-    private Long userId;
-
+    private Integer userId;
     /**
      * 
      */
