+++ src/main/java/edu/hnu/conference_system/service/RoomService.java
@@ -45,7 +45,7 @@ public interface RoomService {
 
     Result joinFromSchedule(String meetingNumber);
 
-    List<Long> getOnMeetingUserId(Long meetingId);
+    List<Integer> getOnMeetingUserId(Long meetingId);
 
     String getAudioPath();
